requirements.txt
fastapi
uvicorn[standard]
python-multipart
moviepy
openai-whisper
transformers
torch
jinja2
